requirements.txt
# Stable versions of the libraries 
py7zr==1.0.0
pyzipper==0.3.6
paramiko==3.5.1 
pycryptodome==3.20.0

# You can use the latest available versions, but be aware that this may lead to dependency errors
